requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
confusable-homoglyphs==3.2.0
Django==3.1a1
django-cors-headers==3.6.0
django-crispy-forms==1.9.2
django-durationwidget==1.0.5
django-filter==2.2.0
django-registration==3.1.1
django-sb-admin==5.5.1.1
djangorestframework==3.11.0
importlib-metadata==1.6.0
jsonfield==3.1.0
Markdown==3.2.2
PyMuPDF==1.22.5
pyserial==3.5
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
zipp==3.1.0
